DOCSP-26099: cs is an async iterable (#498)
* DOCSP-26099: change stream async iterable

* addtl work

* MW PR fixes 1

* small fixes

* MW PR fixes 2

* small fixes

* remove redundant note

(cherry picked from commit 90868281a8144dfbfe047c9f97f09ece1ae63a3d)
